fix: triggers and steps icons alignment
