+++ src/platform/plugins/shared/workflows_management/public/widgets/worflows_triggers_list/popover_items/index.tsx
@@ -102,6 +102,9 @@ const PopoverItemsComponent = <T extends unknown>({
         ownFocus
         data-test-subj={`${dataTestPrefix}DisplayPopover`}
         aria-label={popoverTitle ?? popoverButtonTitle}
+        css={css`
+          display: inline-flex;
+        `}
         button={
           <EuiBadge
             iconType={popoverButtonIcon}
@@ -122,7 +125,7 @@ const PopoverItemsComponent = <T extends unknown>({
             {popoverTitle}
           </EuiPopoverTitle>
         ) : null}
-        <PopoverWrapper data-test-subj={`${dataTestPrefix}DisplayPopoverWrapper`}>
+        <PopoverWrapper data-test-subj={`${dataTestPrefix}DisplayPopoverWrapper`} gutterSize="s">
           <OverflowList items={items.slice(numberOfItemsToDisplay)} />
         </PopoverWrapper>
       </EuiPopover>
